refactor query addon package
